完成了插入用户信息功能。 改了bug：异常处理类初始化时候没有new就直接赋值导致空引用异常、统一异常处理的类没有加@ResponseBody导致返回的不是页面而出现错误

+++ src/main/java/com/liren/lottery_system/common/pojo/dto/RegisterResponseDTO.java
@@ -1,8 +1,6 @@
 package com.liren.lottery_system.common.pojo.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.io.Serializable;
 
@@ -12,5 +10,5 @@ import java.io.Serializable;
 public class RegisterResponseDTO implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private Integer userId;
+    private long userId;
 }
